+++ BaseDeProjetos/Models/Empresa.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using BaseDeProjetos.Models.Enums;
 
 namespace BaseDeProjetos.Models
 {
