+++ BaseDeProjetos/Models/Enums/Base_enums.cs
@@ -1,0 +1,92 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+using System.Reflection;
+
+namespace BaseDeProjetos.Models.Enums
+{
+ 
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    public static class EnumExtensions
+    {
+        public static string GetDisplayName(this Enum enu)
+        {
+            DisplayAttribute attr = GetDisplayAttribute(enu);
+            return attr != null ? attr.Name : enu.ToString();
+        }
+
+        public static string GetDescription(this Enum enu)
+        {
+            DisplayAttribute attr = GetDisplayAttribute(enu);
+            return attr != null ? attr.Description : enu.ToString();
+        }
+
+        private static DisplayAttribute GetDisplayAttribute(object value)
+        {
+            Type type = value.GetType();
+            if (!type.IsEnum)
+            {
+                throw new ArgumentException(string.Format("Type {0} is not an enum", type));
+            }
+
+            // Get the enum field.
+            FieldInfo field = type.GetField(value.ToString());
+            return field?.GetCustomAttribute<DisplayAttribute>();
+        }
+    }
+}
