+++ BaseDeProjetos/Models/Enums/CasaFirjan.cs
@@ -1,0 +1,15 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BaseDeProjetos.Models.Enums
+{
+    public enum CasaFirjan
+    {
+        Firjan,
+
+        [Display(Name = "Firjan SESI")]
+        FirjanSESI,
+
+        [Display(Name = "Firjan SENAI")]
+        FirjanSENAI,
+    }
+}
