+++ BaseDeProjetos/Models/Enums/Idioma.cs
@@ -1,0 +1,8 @@
+﻿namespace BaseDeProjetos.Models.Enums
+{
+    public enum Idioma
+    {
+        ptBR,
+        enUS
+    }
+}
