+++ BaseDeProjetos/Models/Enums/MeioDivulgacao.cs
@@ -1,0 +1,8 @@
+﻿namespace BaseDeProjetos.Models.Enums
+{
+    public enum MeioDivulgacao
+    {
+        //TODO: Capitalizar Fonte
+        IMPRESSO, WEB, MEIO_MAGNETICO, MEIO_DIGITAL, FILME, HIPERTEXTO, OUTRO, VARIOS, NAO_INFORMADO
+    }
+}
