+++ BaseDeProjetos/Models/Enums/Pais.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BaseDeProjetos.Models.Enums
+{
+    public enum Pais
+    {
+        [Display(Name = "Brasil")]
+        Brasil,
+
+        [Display(Name = "Exterior")]
+        Exterior
+    }
+}
