+++ BaseDeProjetos/Models/Enums/SegmentoEmpresa.cs
@@ -1,0 +1,133 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BaseDeProjetos.Models.Enums
+{
+    public enum SegmentoEmpresa
+    {
+        [Display(Name = "Abastecimento de água")]
+        AbastecimentoAgua,
+
+        [Display(Name = "Agro")]
+        Agro,
+
+        [Display(Name = "Alimentos e Bebidas")]
+        AlimentosBebidas,
+
+        [Display(Name = "Automação")]
+        Automacao,
+
+        [Display(Name = "Automobilística")]
+        Automobilistica,
+
+        [Display(Name = "Biomol")]
+        Biomol,
+
+        [Display(Name = "Biorrefinaria")]
+        Biorrefinaria,
+
+        [Display(Name = "Biotecnologia")]
+        Biotecnologia,
+
+        [Display(Name = "Construtora")]
+        Constutora,
+
+        [Display(Name = "Cosméticos")]
+        Cosmeticos,
+
+        [Display(Name = "Embalagens")]
+        Embalagens,
+
+        [Display(Name = "Energia")]
+        Energia,
+
+        [Display(Name = "Equipamentos")]
+        Equipamentos,
+
+        [Display(Name = "Extrativos naturais")]
+        Extrativos,
+
+        [Display(Name = "Farmacêutica")]
+        Farmaceutica,
+
+        [Display(Name = "Formulações")]
+        Formulacoes,
+
+        [Display(Name = "Gases")]
+        Gases,
+
+        [Display(Name = "Lubrificantes")]
+        Lubrificantes,
+
+        [Display(Name = "Madeira")]
+        Madeira,
+
+        [Display(Name = "Mantas absorvedoras")]
+        Mantas,
+
+        [Display(Name = "Materiais")]
+        Materiais,
+
+        [Display(Name = "Mineiração")]
+        Mineiracao,
+
+        [Display(Name = "Nanotecnologia")]
+        Nanotec,
+
+        [Display(Name = "Outros")]
+        Outros,
+
+        [Display(Name = "Papel e celulose")]
+        PapelCelulose,
+
+        [Display(Name = "Petróleo e Gás")]
+        PetroleoGas,
+
+        [Display(Name = "Petroquímica")]
+        Petroquimica,
+
+        [Display(Name = "Polímeros")]
+        Polimeros,
+
+        [Display(Name = "Produção de Sal")]
+        ProducaoSal,
+
+        [Display(Name = "Produtos naturais")]
+        ProdutosNaturais,
+
+        [Display(Name = "Química")]
+        Quimica,
+
+        [Display(Name = "Química 4.0")]
+        Quimica4,
+
+        [Display(Name = "Remediação")]
+        Remediacao,
+
+        [Display(Name = "Saneamento")]
+        Saneamento,
+
+        [Display(Name = "Saúde")]
+        Saude,
+
+        [Display(Name = "Siderurgia/Metalurgia")]
+        SiderurgiaMetalurgia,
+
+        [Display(Name = "Sucroenergético")]
+        Sucroenergetico,
+
+        [Display(Name = "Tecidos")]
+        Tecidos,
+
+        [Display(Name = "Tecnologia")]
+        Tecnologia,
+
+        [Display(Name = "Tintas")]
+        Tintas,
+
+        [Display(Name = "Tratamento de água")]
+        TratamentAgua,
+
+        [Display(Name = "Tratamento de Efluentes")]
+        TratamentEfluentes,
+    }
+}
