+++ BaseDeProjetos/Models/Enums/SetoresDeAtividade.cs
@@ -1,0 +1,93 @@
+﻿namespace BaseDeProjetos.Models.Enums
+{
+    public enum SetoresDeAtividade
+    {
+        Administração_pública, _defesa_e_seguridade_social,
+        Agências_de_viagens, _operadores_turísticos_e_serviços_de_reservas,
+        Agricultura, _Pecuária_e_Serviços_Relacionados,
+        Alimentação,
+        Alojamento,
+        Aluguéis_não_imobiliários_e_gestão_de_ativos_intangíveis_não_financeiros,
+        Armazenamento_e_atividades_auxiliares_dos_transportes,
+        Atividades_artísticas, _criativas_e_de_espetáculos,
+        Atividades_auxiliares_dos_serviços_financeiros_seguros, _previdência_complementar_e_planos_de_saúde,
+        Atividades_cinematográficas, _produção_de_vídeos_e_de_programas_de_televisão_gravação_de_som_e_edição_de_música,
+        Atividades_de_Apoio_à_Extração_de_Minerais,
+        Atividades_de_atenção_à_saúde_humana,
+        Atividades_de_atenção_à_saúde_humana_integradas_com_assistência_social, _prestadas_em_residências_coletivas_e_particulares,
+        Atividades_de_exploração_de_jogos_de_azar_e_apostas,
+        Atividades_de_organizações_associativas,
+        Atividades_de_prestação_de_serviços_de_informação,
+        Atividades_de_rádio_e_de_televisão,
+        Atividades_de_sedes_de_empresas_e_de_consultoria_em_gestão_empresarial,
+        Atividades_de_serviços_financeiros,
+        Atividades_de_vigilância, _segurança_e_investigação,
+        Atividades_dos_serviços_de_tecnologia_da_informação,
+        Atividades_esportivas_e_de_recreação_e_lazer,
+        Atividades_imobiliárias,
+        Atividades_jurídicas, _de_contabilidade_e_de_auditoria,
+        Atividades_ligadas_ao_patrimônio_cultural_e_ambiental,
+        Atividades_veterinárias,
+        Captação, _tratamento_e_distribuição_de_água,
+        Coleta, _tratamento_e_disposição_de_resíduos_recuperação_de_materiais,
+        Comércio_e_reparação_de_veículos_automotores_e_motocicletas,
+        Comércio_por_atacado_exceto_veículos_automotores_e_motocicletas,
+        Comércio_varejista,
+        Confecção_de_artigos_do_vestuário_e_acessórios,
+        Construção_de_edifícios,
+        Correio_e_outras_atividades_de_entrega,
+        Descontaminação_e_outros_serviços_de_gestão_de_resíduos,
+        Edição_e_edição_integrada_à_impressão,
+        Educação,
+        Eletricidade, _gás_e_outras_utilidades,
+        Esgoto_e_atividades_relacionadas,
+        Extração_de_Carvão_Mineral,
+        Extração_de_Minerais_Metálicos,
+        Extração_de_Minerais_Não_Metálicos,
+        Extração_de_Petróleo_e_Gás_Natural,
+        Fabricação_de_Bebidas,
+        Fabricação_de_celulose_papel_e_produtos_de_papel,
+        Fabricação_de_coque_de_produtos_derivados_do_petróleo_e_de_biocombustíveis,
+        Fabricação_de_equipamentos_de_informática_produtos_eletrônicos_e_ópticos,
+        Fabricação_de_máquinas, _aparelhos_e_materiais_elétricos,
+        Fabricação_de_máquinas_e_equipamentos,
+        Fabricação_de_móveis,
+        Fabricação_de_outros_equipamentos_de_transporte_exceto_veículos_automotores,
+        Fabricação_de_Produtos_Alimentícios,
+        Fabricação_de_produtos_de_borracha_e_de_material_plástico,
+        Fabricação_de_produtos_de_madeira,
+        Fabricação_de_produtos_de_metal_exceto_máquinas_e_equipamentos,
+        Fabricação_de_produtos_de_minerais_não_metálicos,
+        Fabricação_de_produtos_diversos,
+        Fabricação_de_Produtos_do_Fumo,
+        Fabricação_de_produtos_farmoquímicos_e_farmacêuticos,
+        Fabricação_de_produtos_químicos,
+        Fabricação_de_Produtos_Têxteis,
+        Fabricação_de_veículos_automotores, _reboques_e_carrocerias,
+        Impressão_e_reprodução_de_gravações,
+        Manutenção, _reparação_e_instalação_de_máquinas_e_equipamentos,
+        Metalurgia,
+        Obras_de_infra_estrutura,
+        Organismos_internacionais_e_outras_instituições_extraterritoriais,
+        Outras_atividades_de_serviços_pessoais,
+        Outras_atividades_profissionais, _científicas_e_técnicas,
+        Pesca_e_Aqüicultura,
+        Pesquisa_e_desenvolvimento_científico,
+        Preparação_de_couros_e_fabricação_de_artefatos_de_couro, _artigos_para_viagem_e_calçados,
+        Produção_Florestal,
+        Publicidade_e_pesquisa_de_mercado,
+        Reparação_e_manutenção_de_equipamentos_de_informática_e_comunicação_e_de_objetos_pessoais_e_domésticos,
+        Seguros, _resseguros_previdência_complementar_e_planos_de_saúde,
+        Seleção, _agenciamento_e_locação_de_mão_de_obra,
+        Serviços_de_arquitetura_e_engenharia_testes_e_análises_técnicas,
+        Serviços_de_assistência_social_sem_alojamento,
+        Serviços_de_escritório, _de_apoio_administrativo_e_outros_serviços_prestados_às_empresas,
+        Serviços_domésticos,
+        Serviços_especializados_para_construção,
+        Serviços_para_edifícios_e_atividades_paisagísticas,
+        Telecomunicações,
+        Transporte_aéreo,
+        Transporte_aquaviário,
+        Transporte_terrestre,
+    }
+}
