+++ BaseDeProjetos/Models/Enums/StatusEdital.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BaseDeProjetos.Models.Enums
+{
+    public enum StatusEdital
+    {
+        [Display(Name = "Em aberto")]
+        aberto,
+
+        [Display(Name = "Edital Encerrado")]
+        encerrado
+    }
+}
