+++ BaseDeProjetos/Models/Enums/TipoInovacao.cs
@@ -1,0 +1,8 @@
+﻿namespace BaseDeProjetos.Models.Enums
+{
+    public enum TipoInovacao
+    {
+        Produto,
+        Processo
+    }
+}
