+++ BaseDeProjetos/Models/Enums/TipoProducao.cs
@@ -1,0 +1,27 @@
+﻿namespace BaseDeProjetos.Models.Enums
+{
+    public enum TipoProducao
+    {
+        //_Tipos_de_Produção_Bibliográfica
+        Trabalhos_em_Eventos,
+
+        Artigos_Publicados,
+        Livros_e_Capítulos, //Ex.: Livro e Capitulo
+        Textos_em_Jornais_ou_Revistas,
+        Demais_Tipos_de_Produção_Bibliográfica,
+        Artigos_Aceitos_Para_Publicação,
+
+        //Tipos_de_Produção_Técnica
+        Cultivo,
+
+        Software,
+        Patente,
+        Desenho_Industrial,
+        Marca,
+        Topografia_de_Circuito_Integrado,
+        Produto_Tecnológico,
+        Processos_ou_Técnicas,
+        Trabalho_Técnico,
+        Demais_Tipos_de_Produção_Técnica
+    }
+}
