+++ BaseDeProjetos/Models/Enums/TipoVinculo.cs
@@ -1,0 +1,11 @@
+﻿namespace BaseDeProjetos.Models.Enums
+{
+    public enum TipoVinculo
+    {
+        Visitante,
+        Aluno,
+        JovemAprendiz,
+        Estagiario,
+        Empregado
+    }
+}
