+++ BaseDeProjetos/Models/PesquisasOpiniao.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text.Json;
+using BaseDeProjetos.Models.Enums;
 
 namespace BaseDeProjetos.Models
 {
