+++ BaseDeProjetos/Models/Projeto.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using BaseDeProjetos.Models.Enums;
 
 namespace BaseDeProjetos.Models
 {
