+++ BaseDeProjetos/Models/Prospeccao.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Diagnostics.CodeAnalysis;
+using BaseDeProjetos.Models.Enums;
 
 namespace BaseDeProjetos.Models
 {
